+++ src/main/java/com/example/bookiibookii/domain/tracker/controller/TrackerController.java
@@ -1,9 +1,10 @@
 package com.example.bookiibookii.domain.tracker.controller;
 
 
+
 import com.example.bookiibookii.domain.tracker.controller.docs.TrackerApi;
-import com.example.bookiibookii.domain.tracker.dto.response.TrackerDetailResponse;
-import com.example.bookiibookii.domain.tracker.dto.response.TrackerHistoryResponse;
+import com.example.bookiibookii.domain.tracker.dto.res.TrackerDetailResponse;
+import com.example.bookiibookii.domain.tracker.dto.res.TrackerHistoryResponse;
 import com.example.bookiibookii.domain.tracker.service.TrackerService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
@@ -13,18 +14,6 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 import java.util.List;
-
-//@RestController
-//@RequestMapping("/api/v1/trackers")
-//@RequiredArgsConstructor
-//public class TrackerController implements TrackerApi {
-//    private final TrackerService trackerService;
-//
-//    @GetMapping("/{trackerId}")
-//    public ResponseEntity<TrackerDetailResponse> getTrackerDetail(@PathVariable Long trackerId) {
-//        return ResponseEntity.ok(trackerService.getTrackerDetail(trackerId));
-//    }
-//}
 
 @RestController
 @RequestMapping("/api/groups") // 주소는 groups로 시작!
